requirements.txt
aniso8601==9.0.1
asgiref==3.5.2
autopep8==1.7.0
black==24.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.7
colorama==0.4.6
Django==4.1.1
django-elo-rating==0.0.1
django-graphql-jwt==0.3.4
django-sesame==3.1
exceptiongroup==1.1.3
flake8==7.0.0
graphene==3.1.1
graphene-django==3.0.0
graphql-core==3.2.3
graphql-relay==3.2.0
greenlet==3.0.3
idna==3.4
iniconfig==2.0.0
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.12.1
platformdirs==4.2.0
pluggy==1.3.0
promise==2.3
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pyflakes==3.2.0
PyJWT==2.5.0
pytest==7.4.2
pytest-asyncio==0.23.4
requests==2.28.1
six==1.16.0
SQLAlchemy==2.0.27
sqlparse==0.4.3
text-unidecode==1.3
toml==0.10.2
tomli==2.0.1
typing_extensions==4.10.0
tzdata==2022.4
urllib3==1.26.12
websocket-client==1.5.1
websockets==10.3